+++ app/src/main/java/id/furqoncreative/jetstories/data/source/network/NetworkDataSource.kt
@@ -22,8 +22,8 @@ interface NetworkDataSource {
         token: String,
         file: MultipartBody.Part,
         description: RequestBody,
-        latitude: Float?,
-        longitude: Float?
+        latitude: Double?,
+        longitude: Double?
     ): Flow<AddStoryResponse>
 
     suspend fun getAllStories(
